chore: refactoring modules folder, enhance the layout of plan list
