+++ travel-neko-web/src/app/plans/page.tsx
@@ -1,5 +1,5 @@
 import { prefetchPlanListQuery } from "@/actions/prefetch-plan-list-query";
-import { PlanList } from "@/module/plan-list";
+import { PlanList } from "@/modules/plan-list";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 
 export default async function PlansPage() {
