+++ travel-neko-web/src/components/plan-list-placeholder/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 export default function PlanListPlaceholder() {
   return (
-    <div className="flex flex-col gap-4">
+    <div className="grid h-full w-full grid-flow-row grid-cols-1 gap-4 lg:grid-cols-2">
       {new Array(10).fill(null).map((item, index) => (
         <div className="bg-placeholder flex h-80 w-full flex-col" key={index}>
           <span className="bg-placeholder block w-full flex-1" />
