+++ travel-neko-web/src/modules/plan-detail-loading/index.tsx
